Re-work how the display mode is determined, and allow it to change during runtime for when games change the video mode
Allow the user to quit the game with the ESC key

Resolve some compiler warnings

+++ src/Utils.cpp
@@ -336,7 +336,7 @@ std::string Utils::strToLower(std::string subject) {
 }
 
 
-std::string Utils::padString(const std::string &string, int length) {
+std::string Utils::padString(const std::string &string, unsigned long length) {
     return string.length() >= length ? string : string + std::string(length - string.size(), ' ');
 }
